sidequest... fix the stability of the tagSelector and made it more composable
