+++ src/core/utils/typedMemo.ts
@@ -1,0 +1,3 @@
+import { memo } from "react";
+
+export const typedMemo: <T>(c: T) => T = memo;
