+++ src/pages/monitoring/outlets/resources/manager/resourcesEditor/ResourceTagSelector.tsx
