+++ src/pages/monitoring/outlets/resources/manager/resourcesEditor/layout/tagCategories.ts
